# docs/swagger.yaml
definitions:
  models.Post:
    properties:
      content:
        type: string
      id:
        type: integer
      post_images:
        items:
          $ref: '#/definitions/models.PostImage'
        type: array
      published_at:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      user_id:
        type: integer
      vote_count:
        type: integer
      votes:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.PostImage:
    properties:
      id:
        type: integer
      post_id:
        type: integer
      url:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.Tag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  requests.RegisterRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  requests.StorePostRequest:
    properties:
      content:
        type: string
      post_images:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  requests.UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  results.JSONResult:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  results.TokenResult:
    properties:
      token:
        type: string
    type: object
  results.UserResult:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /api/v1/auth/login:
    post:
      description: Login, authenticating user by email and password.
      parameters:
      - description: the request body to authenticate user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.TokenResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Login
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Fetch authenticated user account data.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.UserResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Fetch authenticated user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      description: Register, create new user account.
      parameters:
      - description: the request body to create a new user account
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Registering a new account
      tags:
      - Auth
  /api/v1/auth/update-password:
    put:
      description: Update user authenticated password.
      parameters:
      - description: the request body to update a password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdatePasswordRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Update account password
      tags:
      - Posts
  /api/v1/posts:
    get:
      description: Get a list of published posts.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Post'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all published posts
      tags:
      - Posts
  /api/v1/posts/:
    post:
      description: Create a new post.
      parameters:
      - description: the request body to create a new post
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.StorePostRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/requests.StorePostRequest'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Create new post
      tags:
      - Posts
  /api/v1/posts/{id}:
    delete:
      description: Delete an existing post by ID, you can only delete your own post.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a post
      tags:
      - Posts
    get:
      description: Get a post by inserting ID in route param (only published post
        will return the result).
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get published post by ID
      tags:
      - Posts
    put:
      description: Update an existing post by ID, you can only update your own post.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: the request body to update a post
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/requests.StorePostRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/requests.StorePostRequest'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Update a post
      tags:
      - Posts
  /api/v1/posts/{id}/my:
    get:
      description: Get user authenticated post by inserting ID in route param.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get authenticated user post by ID
      tags:
      - Posts
  /api/v1/posts/{id}/publish:
    patch:
      description: Publish an existing post by ID, you can only publish your own post.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/results.JSONResult'
      summary: Publish a post
      tags:
      - Posts
...
  /api/v1/posts/my:
    get:
      description: Get a list of authenticated user posts.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Post'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all authenticated user posts
      tags:
      - Posts
...
swagger: "2.0"